Add example of Cloneable implementation
Example demonstrates how cloning can be performed. Reader should
take not that int is primitive type and copying it will always
result in independent value and String in Java is immutable.
Immutability of String means that it cannot be altered and simple
assignemtn is safe for it. In case of class components being a
mutable objects, user should clone them instead of making simple
reference copy.